handle save event from parent (just log event on parent for the first)
issue #2

In React, we should create as many as dumb components because they
can be reusable component in this case. So, just put save event handler
to parent component (TodoApp) and log the text input. (Then, we need
to save this text input to store via Relay!)

Check console in inspector to see how it works. You can see two log.
One is under TodoTextInput when you type something, another is under
TodoApp, when you hit the enter key.

But, you can see that, input box has text input but users may expect
it should be empty. What's wrong with it?

We will see in the next commit!