add entity via graphql and relay
issue #2

This is really complicated part. So, please take your time, grab
some coffee or tea or beer(NOPE!).

We might need to understand mutation.

https://facebook.github.io/relay/docs/mutations.html

For adding mutation, what we need to do is:

a. create mutations to include all the mutation parts
b. create mutation file and define how mutation works

This mutation should be conformed to the schema.graphql under data,
because this is GraphQL.

Take a time to understand.

What I found is there are tons of conventions of GraphQL and Relay.
So, you might need to not only understand them but also remember
them, because they are convention.

Notice: Please remember, after you change graphQL part, please
rebuild it to see the implementation based on up to date GraphQL
schema.

Also, this is not including optimisticUpdater because that requires
complicated part but we don't need it at this moment.

https://goo.gl/KPHhEa